Fixed edl edit timebar 'f' and 'a' keys to work with multiple reels. Fixed edl edit to work with 3 or more reels.
Former-commit-id: 1e9d305484cd79b2b786d73dbd0298d8eea31c20

+++ mrViewer/src/gui/mrvEDLGroup.h
@@ -25,8 +25,9 @@ class EDLGroup : public fltk::Group
      EDLGroup(int x, int y, int w, int h);
      ~EDLGroup();
 
-     // Add a media track and return its index
-     size_t add_media_track( size_t reel_idx );
+     // Add a media track and return its index.
+     // index can be negative 1 to indicate a missing reel
+     size_t add_media_track( int reel_idx );
 
      // Add an audio only track and return its index
      size_t add_audio_track();
